src/app/confirm-dialog/confirm-dialog.component.ts
import { Component, Inject, OnInit } from '@angular/core';
import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';

@Component({
  selector: 'app-confirm-dialog',
  templateUrl: './confirm-dialog.component.html',
  styleUrls: ['./confirm-dialog.component.scss']
})
export class ConfirmDialogComponent implements OnInit {

  title = 'Confirm';
  confirmMessage: string;
  constructor(public dialogRef: MatDialogRef<ConfirmDialogComponent>,
    @Inject(MAT_DIALOG_DATA) public data: any) { }

  ngOnInit(): void {
    this.confirmMessage = this.data.confirmMessage;
  }

  onConfirm() {   
    this.dialogRef.close(true);
  }

  onCancel() {
    this.dialogRef.close();
  }

}
